fix: type에러 수정.

+++ src/main/java/kr/co/itid/cms/mapper/cms/core/menu/MenuMapper.java
@@ -46,7 +46,7 @@ public interface MenuMapper {
         return MenuTreeResponse.builder()
                 .id(menu.getId())
                 .parentId(menu.getParentId())
-                .position(menu.getPosition().intValue())
+                .position(menu.getPosition())
                 .level(menu.getLevel().intValue())
                 .title(menu.getTitle())
                 .name(menu.getName())
